// src/api/generated/models/addressCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Creates an independent address record owned by another aggregate.
 */
export interface AddressCreate {
  /**
   * @minLength 2
   * @maxLength 150
   */
  street: string
  /**
   * @minLength 1
   * @maxLength 20
   */
  house_number: string
  /**
   * @minLength 4
   * @maxLength 10
   */
  zip_code: string
  /**
   * @minLength 2
   * @maxLength 100
   */
  city: string
  latitude?: number | null
  longitude?: number | null
}